pyproject.toml
[project]
name = "example-flask-auth"
version = "0.0.0"
description = "Flask application demonstrating Zitadel authentication with PKCE using Authlib."
requires-python = ">=3.9"
dependencies = [
  "Flask>=3.0,<4.0",
  "Authlib>=1.3.2,<2.0.0",
  "python-dotenv==1.1.1",
  "requests>=2.32.0,<3.0.0",
]

[project.urls]
repository = "https://github.com/zitadel/example-flask-auth"

[tool.poetry]
name = "example-flask-auth"
version = "0.0.0"
description = "Flask application demonstrating Zitadel authentication with PKCE using Authlib."
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Flask = ">=3.0,<4.0"
Authlib = ">=1.3.2,<2.0.0"
python-dotenv = "1.1.1"
requests = ">=2.32.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.2.1"
pytest-cov = ">=2.8.1"
mypy = ">=1.5"
python-dotenv = "1.1.1"
ruff = "^0.12.4"
fawltydeps = "0.19.0"
types-requests = "^2.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["test", "spec"]
python_files = "test_*.py *_spec.py"
addopts = [
  "--cov=app",
  "--cov=lib",
  "--cov-report=html:build/coverage/html",
  "--cov-report=xml:build/coverage/clover.xml",
]
junit_family = "legacy"

[tool.coverage.run]
data_file = "build/coverage/.coverage"
omit = []

[tool.mypy]
files = ["app", "lib", "test", "spec"]
strict = true
implicit_reexport = true

[[tool.mypy.overrides]]
module = [
  "authlib",
  "authlib.integrations.requests_client",
  "authlib.jose",
]
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[tool.ruff]
line-length = 130
fix = true
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C", "N", "Q", "S", "T"]
extend-select = ["A"]
ignore = ["S101"]

[tool.fawltydeps]
ignore_unused = [
  "fawltydeps",
  "mypy",
  "pytest-cov",
  "ruff",
  "requests",
]
